requirements.txt
selenium
selenium-stealth
webdriver-manager
python-dotenv
hydra-core
omegaconf
